REFACTOR: transient props로 수정 및  폴더 구조 변경
- dom에 불필요하게 생성되는 속성을 제거하기 위해 transient props로 수정
- 폴더 구조 변경
