+++ src/common/components/index.ts
@@ -1,3 +1,6 @@
-import IconButton from './IconButton';
+import IconButton from './IconButton/IconButton';
+import Link from './Link/Link';
+import SearchInput from './SearchInput/SearchInput';
+import SquareButton from './SquareButton/SquareButton';
 
-export { IconButton };
+export { IconButton, Link, SearchInput, SquareButton };
