+++ src/pages/MyCollections/components/index.ts
@@ -1,3 +1,3 @@
-import Bookshelf from './Bookshelf';
+import Bookshelf from './Bookshelf/Bookshelf';
 
 export { Bookshelf };
